mv: map page to main page
